fix draw results
